New authentication settings page
Allows users to select which tenant to authenticate against.

When a user's AAD credentials are associated with multiple tenants, a failed tenant will not prevent users from using BatchExplorer on authorized tenants. When a non-home tenant authorization fails, a notification is displayed asking the user to configure their authentication settings. The authentication settings page displays a list of associated tenants and allows users to deactivate any and to reauthenticate newly activated ones.

Fixes #2161